mudança: alterações na entidade e relações de convite, todos os endpoints dessa entidade estão funcionando
# Conflicts:
#	Back-end/techbridge/src/main/kotlin/com/techbridge/techbridge/entity/Usuario.kt
#	Back-end/techbridge/src/main/kotlin/com/techbridge/techbridge/repository/UsuarioRepository.kt
#	Back-end/techbridge/src/main/resources/application.properties

+++ Back-end/techbridge/src/main/kotlin/com/techbridge/techbridge/dto/AgendaRequestDTO.kt
@@ -6,7 +6,7 @@ import java.time.LocalDateTime
 
 data class AgendaRequestDTO(
     val dataDisponivel: String,
-    val fkGuia: Int?
+    val fkGuia: Long?
 ) {
     fun toEntity(guia: Usuario) = AgendaResponsavel(
         dataDisponivel = LocalDateTime.parse(dataDisponivel),
